Changing naming from "2D" to "SE2"
