+++ include/steam.hpp
@@ -45,5 +45,5 @@
 #include "steam/trajectory/bspline/interface.hpp"
 #include "steam/trajectory/const_acc/interface.hpp"
 #include "steam/trajectory/const_vel/interface.hpp"
-#include "steam/trajectory/const_vel_2d/interface.hpp"
+#include "steam/trajectory/const_vel_se2/interface.hpp"
 #include "steam/trajectory/singer/interface.hpp"
