+++ include/steam/trajectory/const_vel_se2/pose_interpolator.hpp
@@ -5,26 +5,26 @@
 #include "lgmath.hpp"
 
 #include "steam/evaluable/evaluable.hpp"
-#include "steam/trajectory/const_vel_2d/variable.hpp"
+#include "steam/trajectory/const_vel_se2/variable.hpp"
 #include "steam/trajectory/time.hpp"
 
 namespace steam {
 namespace traj {
-namespace const_vel_2d {
+namespace const_vel_se2 {
 
-class VelocityInterpolator : public Evaluable<Eigen::Matrix<double, 3, 1>> {
+class PoseInterpolator : public Evaluable<lgmath::se2::Transformation> {
  public:
-  using Ptr = std::shared_ptr<VelocityInterpolator>;
-  using ConstPtr = std::shared_ptr<const VelocityInterpolator>;
+  using Ptr = std::shared_ptr<PoseInterpolator>;
+  using ConstPtr = std::shared_ptr<const PoseInterpolator>;
 
   using InPoseType = lgmath::se2::Transformation;
   using InVelType = Eigen::Matrix<double, 3, 1>;
-  using OutType = Eigen::Matrix<double, 3, 1>;
+  using OutType = lgmath::se2::Transformation;
 
   static Ptr MakeShared(const Time time, const Variable::ConstPtr& knot1,
                         const Variable::ConstPtr& knot2);
-  VelocityInterpolator(const Time time, const Variable::ConstPtr& knot1,
-                       const Variable::ConstPtr& knot2);
+  PoseInterpolator(const Time time, const Variable::ConstPtr& knot1,
+                   const Variable::ConstPtr& knot2);
 
   bool active() const override;
   void getRelatedVarKeys(KeySet& keys) const override;
@@ -44,6 +44,6 @@ class VelocityInterpolator : public Evaluable<Eigen::Matrix<double, 3, 1>> {
       lambda22_;
 };
 
-}  // namespace const_vel_2d
+}  // namespace const_vel_se2
 }  // namespace traj
 }  // namespace steam
